benchrunner: don't pre-allocate all arrays at once (fix #41)
