+++ benchrunner/MVector.hs
@@ -1,0 +1,9 @@
+-- | Reexport of the right kind of mutable vectors
+
+module MVector
+  (
+    module Data.Vector.Unboxed.Mutable
+  )
+  where
+
+import Data.Vector.Unboxed.Mutable
