+++ benchrunner/Measure.hs
@@ -1,4 +1,4 @@
-module Measure (benchAndRunCSorts, benchAndRunCxxSorts, benchAndRunDataVecSorts, benchOnArrays, bench, benchPar, dotrialIO, benchIO, benchParIO) where
+module Measure where
 
 import Control.Exception (evaluate)
 import Control.Monad.Par hiding (runParIO)
@@ -8,20 +8,13 @@ import Data.Int
 import System.Mem (performMajorGC)
 import Data.Time.Clock (getCurrentTime, diffUTCTime)
 
-import Foreign as F
 import qualified Array as A
-import Types as T (SortAlgo(..), Vec, VecSort)
-import qualified Data.List as L
-import qualified Data.Vector.Unboxed as V
-import qualified ForeignFunctionImports as FFI
-import qualified Data.Vector.Unboxed.Mutable as MV
+import Foreign as F
+import Sort
+import Utils
+import qualified Vector as V
+import qualified MVector as MV
 
---------------------------------------------------------------------------------
-
-median :: [Double] -> Double
-median ls = (L.sort ls) !! (length ls `div` 2)
-
---------------------------------------------------------------------------------
 
 benchPar :: (NFData a, NFData b) =>
             (a -> Par b) -> a -> Int -> IO (b, Double, Double)
@@ -56,7 +49,6 @@ benchIO f arg iters = do
     let  selftimed = median times
          batchtime = sum times
     return $! (last results, selftimed, batchtime)
-
 
 {-# NOINLINE dotrialPar #-}
 dotrialPar :: (NFData a, NFData b) =>
@@ -96,24 +88,6 @@ dotrialIO f arg = do
 
 --------------------------------------------------------------------------------
 
-bench :: (NFData a, Show b, NFData b) => (a %p -> b) -> a -> Int -> IO (b, Double, Double)
-bench f arg iters = do
-    let !arg2 = force arg
-    !tups <- mapM (\_ -> dotrial f arg2) [1..iters]
-    let (results, times) = unzip tups
-    let selftimed = median times
-        batchtime = sum times
-    return $! (last results, selftimed, batchtime)
-
-benchOnArrays :: (NFData a, Show b, NFData b, Show a) => (A.Array a %p -> b) -> [A.Array a] -> IO (b, Double, Double)
-benchOnArrays f arrArgs = do
-    let !arg2s = force arrArgs
-    !tups <- mapM (\arg2' -> dotrial f (force arg2')) arg2s
-    let (results, times) = unzip tups
-    let selftimed = median times
-        batchtime = sum times
-    return $! (last results, selftimed, batchtime)
-
 {-# NOINLINE dotrial #-}
 dotrial :: (NFData a, Show b, NFData b) => (a %p -> b) -> a -> IO (b, Double)
 dotrial f arg = do
@@ -125,13 +99,40 @@ dotrial f arg = do
     putStrLn ("iter time: " ++ show delt)
     return $! (a,delt)
 
+
+bench :: (NFData a, Show b, NFData b) => (a %p -> b) -> a -> Int -> IO (b, Double, Double)
+bench f arg iters = do
+    let !arg2 = force arg
+    !tups <- mapM (\_ -> dotrial f arg2) [1..iters]
+    let (results, times) = unzip tups
+    let selftimed = median times
+        batchtime = sum times
+    return $! (last results, selftimed, batchtime)
+
+benchOnArrays ::
+  (NFData a, Show b, NFData b, Show a, A.HasPrim a) =>
+  (A.Array a %p -> b) -> A.Array a -> Int -> IO (b, Double, Double)
+benchOnArrays f arr iters = do
+  let go (i, a)
+        | i == 0 = pure Nothing
+        | otherwise = do
+            !b <- copyArrayInplaceIO arr a
+            res <- dotrial f b
+            pure $ Just (res, (i - 1, b))
+  !tups <- unfoldrM go (iters, A.make (A.size arr) (A.get arr 0))
+
+  let (results, times) = unzip tups
+      selftimed = median times
+      batchtime = sum times
+  pure (last results, selftimed, batchtime)
+
 benchAndRunDataVecSorts :: VecSort -> Vec -> Int ->  IO (Vec, Double, Double)
-benchAndRunDataVecSorts sortFn inVec iters = do
+benchAndRunDataVecSorts sortfn inVec iters = do
   !tups <- mapM (\_ -> do
                        mvec <- V.thaw inVec
                        mvecCopy <- MV.new (MV.length mvec)
                        MV.copy mvecCopy mvec
-                       res <- dotrialLocal sortFn mvecCopy
+                       res <- dotrialLocal sortfn mvecCopy
                        pure res
                 ) [1..iters]
   let (results, times) = unzip tups
@@ -149,20 +150,6 @@ benchAndRunDataVecSorts sortFn inVec iters = do
       putStrLn ("iter time: " ++ show delt)
       arg' <- V.freeze arg
       return $! (arg', delt)
-
-sortFnC :: SortAlgo -> FFI.SortFn
-sortFnC alg = case alg of
-                    Insertionsort -> FFI.c_insertionsort
-                    Mergesort -> FFI.c_mergesort
-                    Quicksort -> FFI.c_quicksort
-                    _ -> error "sortFnC: Csort not implemented!"
-
-sortFnCxx :: SortAlgo -> FFI.SortFnCxx
-sortFnCxx alg = case alg of
-                    Insertionsort -> FFI.cxx_int_insertionsort
-                    Mergesort -> FFI.cxx_int_mergesort
-                    Quicksort -> FFI.cxx_int_quicksort
-                    _ -> error "sortFnCxx: Csort not implemented!"
 
 -- return type : IO ([Int64], Double, Double)
 -- [Int64]: sorted output array from the last iteration that was run
