+++ benchrunner/Vector.hs
@@ -1,0 +1,9 @@
+-- | Reexport of the right kind of vectors
+
+module Vector
+  (
+    module Data.Vector.Unboxed
+  )
+  where
+
+import Data.Vector.Unboxed
